requirements.txt
django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
requests==2.32.4
drf-yasg==1.21.10
pillow==11.3.0
